fix typo

+++ sort/insertion_sort.h
@@ -4,7 +4,7 @@
 #include <stddef.h>
 
 // insertion sort of values
-void insertion_sort_val(void *array, const unsigned int size, const size_t n, int(*cmp)(const void*, const void*));
+void insertion_sort_val(void *array, const size_t size, const size_t n, int(*cmp)(const void*, const void*));
 
 // insertion sort of pointers
 void insertion_sort_ptr(void **array, const size_t n, int(*cmp)(const void*, const void*));
